Cloud Functions - Generic/requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
datetime
boto3
pandas
pyarrow
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-storage